+++ samza-azure/src/main/java/org/apache/samza/coordinator/AzureLeaderElector.java
@@ -17,12 +17,12 @@
  * under the License.
  */
 
-package org.apache.samza;
+package org.apache.samza.coordinator;
 
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
-import org.apache.samza.coordinator.LeaderElector;
-import org.apache.samza.coordinator.LeaderElectorListener;
+import org.apache.samza.AzureException;
+import org.apache.samza.util.LeaseBlobManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -61,7 +61,7 @@ public class AzureLeaderElector implements LeaderElector {
    * @throws AzureException If a Azure storage service error occurred. This includes the case where the blob you're trying to lease does not exist.
    */
   @Override
-  public void tryBecomeLeader() {
+  public void tryBecomeLeader() throws AzureException {
     leaseId.getAndSet(leaseBlobManager.acquireLease(LEASE_TIME_IN_SEC, leaseId.get()));
     if (leaseId.get() != null) {
       LOG.info("Became leader with lease ID {}.", leaseId.get());
