+++ samza-azure/src/main/java/org/apache/samza/coordinator/data/BarrierState.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.samza;
+package org.apache.samza.coordinator.data;
 
 /**
  * Enum depicting different barrier states.
