+++ samza-azure/src/main/java/org/apache/samza/coordinator/data/JobModelBundle.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.samza;
+package org.apache.samza.coordinator.data;
 
 import org.apache.samza.job.model.JobModel;
