+++ samza-azure/src/main/java/org/apache/samza/coordinator/scheduler/TaskScheduler.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.samza.scheduler;
+package org.apache.samza.coordinator.scheduler;
 
 import java.util.concurrent.ScheduledFuture;
